refactor: move PaymentContent to commonMain
Splits PaymentContent out of PaymentScreen.kt into commonMain so it can
be exercised by PreviewScreenshotGeneratorTest on the JVM target.
